Fix: 리다이렉션 로직 hook 처리
