+++ src/components/common/topBtn.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import TopSign from '../../assets/images/top.svg';
+import BtnForm from '../../assets/images/btn.svg';
+
+const TopBtn = () => {
+	return (
+		<button>
+			<img src={TopSign} alt="Top Sign" />
+		</button>
+	);
+};
+
+export default TopBtn;
